moved the shared Azure authentication helper into common_shared_utils.py and updated the FOCUS utility to reuse it
